Updated Files

+++ backend/controllers/applicationController.js
@@ -85,6 +85,9 @@ export const jobseekerDeleteApplication = catchAsyncErrors(async (req, res, next
   if (!application) {
     return next(new ErrorHandler("Application not found!", 404));
   }
+  if (application.applicantID.user.toString() !== req.user._id.toString()) {
+    return next(new ErrorHandler("You can delete only your own applications.", 403));
+  }
   await application.deleteOne();
   res.status(200).json({
     success: true,
